+++ hijack/hijacker.go
@@ -4,7 +4,7 @@ import (
 	"io"
 	"net"
 
-	"github.com/haxii/fastproxy/header"
+	"github.com/haxii/fastproxy/http"
 )
 
 //Hijacker http hijacker
@@ -12,14 +12,14 @@ type Hijacker interface {
 	// GetRequestWriter return a request sniffing writer based on uri & header
 	GetRequestWriter(
 		host string, method, path []byte,
-		header header.Header, rawHeader []byte) io.Writer
+		header http.Header, rawHeader []byte) io.Writer
 	// GetResponseReader return a response hijacking reader
 	//  a non-nil reader means proxy would stop the request to target
 	//  server then return the reader's response
 	GetResponseReader() io.Reader
 	// GetResponseWriter return a response sniffing writer based on status code & header
 	GetResponseWriter(statusCode int,
-		header header.Header, rawHeader []byte) io.Writer
+		header http.Header, rawHeader []byte) io.Writer
 }
 
 //HijackerPool pooling hijacker instances
