+++ http/header.go
@@ -1,4 +1,4 @@
-package header
+package http
 
 import (
 	"bufio"
