+++ http/status.go
@@ -1,4 +1,4 @@
-package header
+package http
 
 import (
 	"fmt"
